+++ packages/drizzle/migrations/20260827175917_round_the_anarchist/migration.sql
@@ -1,2 +1,3 @@
+-- Custom SQL migration file, put your code below! --
 CREATE INDEX "user_lists_stations_internal_gin" ON "user_lists" USING gin (("stations" -> 'internal'));--> statement-breakpoint
 CREATE INDEX "user_lists_stations_uke_gin" ON "user_lists" USING gin (("stations" -> 'uke'));